% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_mediqr.R
\name{ff_mediqr}
\alias{ff_mediqr}
\title{Parse a binary variable as 'median' and 'interquartile range'}
\usage{
ff_mediqr(var, data = dt, dp = 0)
}
\arguments{
\item{var}{a numerical variable}

\item{data}{a data frame (or data.table)}

\item{dp}{number of decimal places in formatted response}
}
\description{
Return a list that summarises the median and IQR, or mean and SD
}
\examples{
# Create data.frame and summarise proportion of A's and B's
set.seed(1)
dt <- data.frame(choice=sample(c("A", "B"), 100, replace=TRUE), x=rnorm(100))
head(dt)
ff_mediqr(x, data=dt, dp=1)
dd <- data.table(x=rnorm(100,42,6))
ff_msd(x, dd)
}
\keyword{dsbc}
\keyword{rmarkdown}

