% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one.R
\name{table_one}
\alias{table_one}
\title{Prepare Table 1 for biomedical research papers}
\usage{
table_one(dt, table.dict, grp = NULL)
}
\arguments{
\item{dt}{a long data.table with columns for 'id', 'item', and 'val' (and optionally 'group')}

\item{table.dict}{(see notes below)}

\item{grp}{a grouping variable (will produce separate summaries per group)}
}
\description{
Given a long form data.table, then summarise each item, group as necessary,
and prepare a publication quality Table 1
}
\examples{
# table_one <- function(dt, table.dict, grp=NULL) 
}
\keyword{cchic}
\keyword{dsbc}
\keyword{rmarkdown}

