% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookfor.R
\name{lookfor}
\alias{lookfor}
\title{R version of the stata lookfor function}
\usage{
lookfor(string_to_find, data = NULL)
}
\arguments{
\item{string_to_find}{This is the string you wish to find passed. May be quoted or not.}
}
\description{
This will try to find a variable when you can't remember its exact name.
}
\examples{
lookfor_function()
}
\keyword{dsbc}
\keyword{stata,}

