% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookfor.R
\name{lookfor}
\alias{lookfor}
\title{R version of the Stata lookfor function}
\usage{
lookfor(string_to_find, data = dt)
}
\arguments{
\item{string_to_find}{This is the string you wish to find passed. May be quoted or not.}

\item{data}{data frame or table to search}
}
\description{
This will try to find a variable when you can't remember its exact name.
R version of the stata lookfor function
}
\examples{
# Create dt dataframe called dt then find vars with 'y' in the name
dt <- data.frame(id=c(1:10), x=letters[1:10], y1=rnorm(10), y2=round(runif(10)*100))
lookfor(y, data=dt)
}
\keyword{dsbc}
\keyword{stata}

